MAINT: Separate auto generated ufuncs multimethods
