+++ scipy/special/tests/test_orthogonal_eval.py
@@ -2,7 +2,7 @@ import numpy as np
 from numpy.testing import assert_, assert_allclose
 import pytest
 
-from scipy.special import _multimethods
+from scipy.special import _multimethods_ufuncs, _multimethods
 from scipy.special._testutils import FuncData
 
 
@@ -10,23 +10,23 @@ def test_eval_chebyt():
     n = np.arange(0, 10000, 7)
     x = 2*np.random.rand() - 1
     v1 = np.cos(n*np.arccos(x))
-    v2 = _multimethods.eval_chebyt(n, x)
+    v2 = _multimethods_ufuncs.eval_chebyt(n, x)
     assert_(np.allclose(v1, v2, rtol=1e-15))
 
 
 def test_eval_genlaguerre_restriction():
     # check it returns nan for alpha <= -1
-    assert_(np.isnan(_multimethods.eval_genlaguerre(0, -1, 0)))
-    assert_(np.isnan(_multimethods.eval_genlaguerre(0.1, -1, 0)))
+    assert_(np.isnan(_multimethods_ufuncs.eval_genlaguerre(0, -1, 0)))
+    assert_(np.isnan(_multimethods_ufuncs.eval_genlaguerre(0.1, -1, 0)))
 
 
 def test_warnings():
     # ticket 1334
     with np.errstate(all='raise'):
         # these should raise no fp warnings
-        _multimethods.eval_legendre(1, 0)
-        _multimethods.eval_laguerre(1, 1)
-        _multimethods.eval_gegenbauer(1, 1, 0)
+        _multimethods_ufuncs.eval_legendre(1, 0)
+        _multimethods_ufuncs.eval_laguerre(1, 1)
+        _multimethods_ufuncs.eval_gegenbauer(1, 1, 0)
 
 
 class TestPolys:
@@ -69,68 +69,68 @@ class TestPolys:
             ds.check()
 
     def test_jacobi(self):
-        self.check_poly(_multimethods.eval_jacobi, _multimethods.jacobi,
+        self.check_poly(_multimethods_ufuncs.eval_jacobi, _multimethods.jacobi,
                         param_ranges=[(-0.99, 10), (-0.99, 10)],
                         x_range=[-1, 1], rtol=1e-5)
 
     def test_sh_jacobi(self):
-        self.check_poly(_multimethods.eval_sh_jacobi, _multimethods.sh_jacobi,
+        self.check_poly(_multimethods_ufuncs.eval_sh_jacobi, _multimethods.sh_jacobi,
                         param_ranges=[(1, 10), (0, 1)], x_range=[0, 1],
                         rtol=1e-5)
 
     def test_gegenbauer(self):
-        self.check_poly(_multimethods.eval_gegenbauer, _multimethods.gegenbauer,
+        self.check_poly(_multimethods_ufuncs.eval_gegenbauer, _multimethods.gegenbauer,
                         param_ranges=[(-0.499, 10)], x_range=[-1, 1],
                         rtol=1e-7)
 
     def test_chebyt(self):
-        self.check_poly(_multimethods.eval_chebyt, _multimethods.chebyt,
+        self.check_poly(_multimethods_ufuncs.eval_chebyt, _multimethods.chebyt,
                         param_ranges=[], x_range=[-1, 1])
 
     def test_chebyu(self):
-        self.check_poly(_multimethods.eval_chebyu, _multimethods.chebyu,
+        self.check_poly(_multimethods_ufuncs.eval_chebyu, _multimethods.chebyu,
                         param_ranges=[], x_range=[-1, 1])
 
     def test_chebys(self):
-        self.check_poly(_multimethods.eval_chebys, _multimethods.chebys,
+        self.check_poly(_multimethods_ufuncs.eval_chebys, _multimethods.chebys,
                         param_ranges=[], x_range=[-2, 2])
 
     def test_chebyc(self):
-        self.check_poly(_multimethods.eval_chebyc, _multimethods.chebyc,
+        self.check_poly(_multimethods_ufuncs.eval_chebyc, _multimethods.chebyc,
                         param_ranges=[], x_range=[-2, 2])
 
     def test_sh_chebyt(self):
         with np.errstate(all='ignore'):
-            self.check_poly(_multimethods.eval_sh_chebyt, _multimethods.sh_chebyt,
+            self.check_poly(_multimethods_ufuncs.eval_sh_chebyt, _multimethods.sh_chebyt,
                             param_ranges=[], x_range=[0, 1])
 
     def test_sh_chebyu(self):
-        self.check_poly(_multimethods.eval_sh_chebyu, _multimethods.sh_chebyu,
+        self.check_poly(_multimethods_ufuncs.eval_sh_chebyu, _multimethods.sh_chebyu,
                         param_ranges=[], x_range=[0, 1])
 
     def test_legendre(self):
-        self.check_poly(_multimethods.eval_legendre, _multimethods.legendre,
+        self.check_poly(_multimethods_ufuncs.eval_legendre, _multimethods.legendre,
                         param_ranges=[], x_range=[-1, 1])
 
     def test_sh_legendre(self):
         with np.errstate(all='ignore'):
-            self.check_poly(_multimethods.eval_sh_legendre, _multimethods.sh_legendre,
+            self.check_poly(_multimethods_ufuncs.eval_sh_legendre, _multimethods.sh_legendre,
                             param_ranges=[], x_range=[0, 1])
 
     def test_genlaguerre(self):
-        self.check_poly(_multimethods.eval_genlaguerre, _multimethods.genlaguerre,
+        self.check_poly(_multimethods_ufuncs.eval_genlaguerre, _multimethods.genlaguerre,
                         param_ranges=[(-0.99, 10)], x_range=[0, 100])
 
     def test_laguerre(self):
-        self.check_poly(_multimethods.eval_laguerre, _multimethods.laguerre,
+        self.check_poly(_multimethods_ufuncs.eval_laguerre, _multimethods.laguerre,
                         param_ranges=[], x_range=[0, 100])
 
     def test_hermite(self):
-        self.check_poly(_multimethods.eval_hermite, _multimethods.hermite,
+        self.check_poly(_multimethods_ufuncs.eval_hermite, _multimethods.hermite,
                         param_ranges=[], x_range=[-100, 100])
 
     def test_hermitenorm(self):
-        self.check_poly(_multimethods.eval_hermitenorm, _multimethods.hermitenorm,
+        self.check_poly(_multimethods_ufuncs.eval_hermitenorm, _multimethods.hermitenorm,
                         param_ranges=[], x_range=[-100, 100])
 
 
@@ -175,76 +175,76 @@ class TestRecurrence:
             ds.check()
 
     def test_jacobi(self):
-        self.check_poly(_multimethods.eval_jacobi,
+        self.check_poly(_multimethods_ufuncs.eval_jacobi,
                         param_ranges=[(-0.99, 10), (-0.99, 10)],
                         x_range=[-1, 1])
 
     def test_sh_jacobi(self):
-        self.check_poly(_multimethods.eval_sh_jacobi,
+        self.check_poly(_multimethods_ufuncs.eval_sh_jacobi,
                         param_ranges=[(1, 10), (0, 1)], x_range=[0, 1])
 
     def test_gegenbauer(self):
-        self.check_poly(_multimethods.eval_gegenbauer,
+        self.check_poly(_multimethods_ufuncs.eval_gegenbauer,
                         param_ranges=[(-0.499, 10)], x_range=[-1, 1])
 
     def test_chebyt(self):
-        self.check_poly(_multimethods.eval_chebyt,
+        self.check_poly(_multimethods_ufuncs.eval_chebyt,
                         param_ranges=[], x_range=[-1, 1])
 
     def test_chebyu(self):
-        self.check_poly(_multimethods.eval_chebyu,
+        self.check_poly(_multimethods_ufuncs.eval_chebyu,
                         param_ranges=[], x_range=[-1, 1])
 
     def test_chebys(self):
-        self.check_poly(_multimethods.eval_chebys,
+        self.check_poly(_multimethods_ufuncs.eval_chebys,
                         param_ranges=[], x_range=[-2, 2])
 
     def test_chebyc(self):
-        self.check_poly(_multimethods.eval_chebyc,
+        self.check_poly(_multimethods_ufuncs.eval_chebyc,
                         param_ranges=[], x_range=[-2, 2])
 
     def test_sh_chebyt(self):
-        self.check_poly(_multimethods.eval_sh_chebyt,
+        self.check_poly(_multimethods_ufuncs.eval_sh_chebyt,
                         param_ranges=[], x_range=[0, 1])
 
     def test_sh_chebyu(self):
-        self.check_poly(_multimethods.eval_sh_chebyu,
+        self.check_poly(_multimethods_ufuncs.eval_sh_chebyu,
                         param_ranges=[], x_range=[0, 1])
 
     def test_legendre(self):
-        self.check_poly(_multimethods.eval_legendre,
+        self.check_poly(_multimethods_ufuncs.eval_legendre,
                         param_ranges=[], x_range=[-1, 1])
 
     def test_sh_legendre(self):
-        self.check_poly(_multimethods.eval_sh_legendre,
+        self.check_poly(_multimethods_ufuncs.eval_sh_legendre,
                         param_ranges=[], x_range=[0, 1])
 
     def test_genlaguerre(self):
-        self.check_poly(_multimethods.eval_genlaguerre,
+        self.check_poly(_multimethods_ufuncs.eval_genlaguerre,
                         param_ranges=[(-0.99, 10)], x_range=[0, 100])
 
     def test_laguerre(self):
-        self.check_poly(_multimethods.eval_laguerre,
+        self.check_poly(_multimethods_ufuncs.eval_laguerre,
                         param_ranges=[], x_range=[0, 100])
 
     def test_hermite(self):
-        v = _multimethods.eval_hermite(70, 1.0)
+        v = _multimethods_ufuncs.eval_hermite(70, 1.0)
         a = -1.457076485701412e60
         assert_allclose(v, a)
 
 
 def test_hermite_domain():
     # Regression test for gh-11091.
-    assert np.isnan(_multimethods.eval_hermite(-1, 1.0))
-    assert np.isnan(_multimethods.eval_hermitenorm(-1, 1.0))
+    assert np.isnan(_multimethods_ufuncs.eval_hermite(-1, 1.0))
+    assert np.isnan(_multimethods_ufuncs.eval_hermitenorm(-1, 1.0))
 
 
 @pytest.mark.parametrize("n", [0, 1, 2])
 @pytest.mark.parametrize("x", [0, 1, np.nan])
 def test_hermite_nan(n, x):
     # Regression test for gh-11369.
-    assert np.isnan(_multimethods.eval_hermite(n, x)) == np.any(np.isnan([n, x]))
-    assert np.isnan(_multimethods.eval_hermitenorm(n, x)) == np.any(np.isnan([n, x]))
+    assert np.isnan(_multimethods_ufuncs.eval_hermite(n, x)) == np.any(np.isnan([n, x]))
+    assert np.isnan(_multimethods_ufuncs.eval_hermitenorm(n, x)) == np.any(np.isnan([n, x]))
 
 
 @pytest.mark.parametrize('n', [0, 1, 2, 3.2])
@@ -252,7 +252,7 @@ def test_hermite_nan(n, x):
 @pytest.mark.parametrize('x', [2, np.nan])
 def test_genlaguerre_nan(n, alpha, x):
     # Regression test for gh-11361.
-    nan_laguerre = np.isnan(_multimethods.eval_genlaguerre(n, alpha, x))
+    nan_laguerre = np.isnan(_multimethods_ufuncs.eval_genlaguerre(n, alpha, x))
     nan_arg = np.any(np.isnan([n, alpha, x]))
     assert nan_laguerre == nan_arg
 
@@ -262,6 +262,6 @@ def test_genlaguerre_nan(n, alpha, x):
 @pytest.mark.parametrize('x', [1e-6, 2, np.nan])
 def test_gegenbauer_nan(n, alpha, x):
     # Regression test for gh-11370.
-    nan_gegenbauer = np.isnan(_multimethods.eval_gegenbauer(n, alpha, x))
+    nan_gegenbauer = np.isnan(_multimethods_ufuncs.eval_gegenbauer(n, alpha, x))
     nan_arg = np.any(np.isnan([n, alpha, x]))
     assert nan_gegenbauer == nan_arg
